Did a clean build, leading to a bunch of bug-fixes.  Also added constructor definitions and added the capability to define both explicitly-coded and default constructors for sum types.

+++ src/jllvm/llvm/AnalysisJNI.java
@@ -9,8 +9,8 @@
 package jllvm.llvm;
 
 class AnalysisJNI {
-  public final static native long LLVMVerifyModule(long jarg1, int jarg2, long jarg3);
-  public final static native long LLVMVerifyFunction(long jarg1, int jarg2);
+  public final static native int LLVMVerifyModule(long jarg1, int jarg2, long jarg3);
+  public final static native int LLVMVerifyFunction(long jarg1, int jarg2);
   public final static native void LLVMViewFunctionCFG(long jarg1);
   public final static native void LLVMViewFunctionCFGOnly(long jarg1);
 }
